+++ bmcd/src/app/bmc_application.rs
@@ -433,7 +433,7 @@ impl BmcApplication {
 
         self.reboot_into_usb(node, UsbConfig::Flashing(node, UsbRoute::Bmc))
             .await?;
-        let blk_dev = self.node_drivers.load_as_block_device().await?;
+        let blk_dev = self.node_drivers.load_as_block_device(node).await?;
 
         if let Err(e) = append_msd_config_to_usb_gadget(&blk_dev).await {
             tracing::error!("msd usb-gadget: {:#}", e);
@@ -451,7 +451,7 @@ impl BmcApplication {
     ) -> anyhow::Result<impl 'static + AsyncRead + AsyncWrite + AsyncSeek + Unpin> {
         self.reboot_into_usb(node, UsbConfig::Flashing(node, router))
             .await?;
-        Ok(self.node_drivers.load_as_stream().await?)
+        Ok(self.node_drivers.load_as_stream(node).await?)
     }
 
     async fn reboot_into_usb(&self, node: NodeId, config: UsbConfig) -> anyhow::Result<()> {